requirements.txt
discord.py
jishaku
PyNaCl
asyncio
requests